is_admin

+++ MisakiAobaBot.py
@@ -130,7 +130,8 @@ def start(bot, update):
 def help(bot, update):
     """Send a message when the command /help is issued."""
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         if randrange(1000)<30:
             bot.send_message(chat_id=update.message.chat_id, text="ぜ")
         else:
@@ -140,7 +141,8 @@ def help(bot, update):
 def tbgame(bot, update):
     """Send a message when the command /tbgame is issued."""
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_tbgame, 
                         parse_mode=ParseMode.HTML)
 
@@ -148,7 +150,8 @@ def tbgame(bot, update):
 def rule(bot, update):
     """Send a message when the command /rule is issued."""
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         if randrange(1000)<30:
             bot.send_message(chat_id=update.message.chat_id, text="ぜ")
         else:
@@ -172,7 +175,8 @@ def config(bot, update, args):
         if not args:
             bot.send_message(chat_id=update.message.chat_id, text="本功能目前沒有毛用")
         else:
-            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+            if is_admin(bot,update):
+                bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
             msg_1=bot.send_message(chat_id=update.message.chat_id, text=word_kachikoi_name,
             parse_mode=ParseMode.HTML)
             time.sleep(6)
@@ -190,7 +194,8 @@ def config(bot, update, args):
 def nanto(bot, update, args):
     """Send a message when the command /nanto is issued."""
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         if not args:
             msg_1=bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_nanto_1)
             time.sleep(1)
@@ -244,7 +249,6 @@ def which(bot, update, args):
             text="わたしは〜♬［$res］が良いと思うよ〜えへへ。".replace('$res',result)
             msg=bot.send_message(chat_id=update.message.chat_id, text=text)
 
-
 def dice(bot,update,args):
     """Send a message when the command /dice is issued."""
     dice=['⚀','⚁','⚂','⚃','⚄','⚅']
@@ -295,7 +299,8 @@ def tiger(bot, update):
     word_tiger_11="<pre>ジャージャー！</pre>"
     word_tiger_12="<pre>ファイボー！ワイパー！</pre>"
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         messg = bot.send_message(chat_id=update.message.chat_id, text=word_tiger_1,
             parse_mode=ParseMode.HTML)
         time.sleep(0.5)
@@ -396,7 +401,8 @@ def set_remind_time(bot,update,args):
 def notiger(bot, update):
     """Send a message when the command /notiger is issued."""
     if update.message.date > init_time:
-        bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
+        if is_admin(bot,update):
+            bot.delete_message(chat_id=update.message.chat_id, message_id=update.message.message_id)
         msg=bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_notiger, 
                     parse_mode=ParseMode.HTML)
         time.sleep(10)
